seperate test for keccak circuit
