+++ zkevm-circuits/src/keccak_circuit.rs
@@ -1,7 +1,11 @@
 //! The keccak circuit implementation.
 /// Keccak packed multi
 pub mod keccak_packed_multi;
+/// test
+#[cfg(any(feature = "test", test))]
+mod test;
 /// Util
 pub mod util;
 
 pub use keccak_packed_multi::KeccakCircuitConfig as KeccakConfig;
+pub use test::KeccakCircuit as TestKeccakCircuit;
